Moving all cross mwm implementation to a single class CrossMwmIndexGraph
