+++ routing/CMakeLists.txt
@@ -22,9 +22,8 @@ set(
   car_model.hpp
   car_router.cpp
   car_router.hpp
+  cross_mwm_index_graph.cpp
   cross_mwm_index_graph.hpp
-  cross_mwm_index_graph_osrm.cpp
-  cross_mwm_index_graph_osrm.hpp
   cross_mwm_road_graph.cpp
   cross_mwm_road_graph.hpp
   cross_mwm_router.cpp
