+++ routing/routing.pro
@@ -19,7 +19,7 @@ SOURCES += \
     bicycle_model.cpp \
     car_model.cpp \
     car_router.cpp \
-    cross_mwm_index_graph_osrm.cpp \
+    cross_mwm_index_graph.cpp \
     cross_mwm_road_graph.cpp \
     cross_mwm_router.cpp \
     cross_routing_context.cpp \
@@ -73,7 +73,6 @@ HEADERS += \
     car_model.hpp \
     car_router.hpp \
     cross_mwm_index_graph.hpp \
-    cross_mwm_index_graph_osrm.hpp \
     cross_mwm_road_graph.hpp \
     cross_mwm_router.hpp \
     cross_routing_context.hpp \
